% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMMarith.R
\name{ADMMarith}
\alias{ADMMarith}
\alias{vector.2.norm}
\alias{shrinkage}
\title{Basic math functions for ADMM}
\usage{
vector.2.norm(x)

shrinkage(x, kappa)
}
\arguments{
\item{x}{numeric vector}

\item{kappa}{shrinkage factor}
}
\description{
Basic math functions for ADMM
}
