% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLasso.R
\docType{package}
\name{ECLasso-package}
\alias{ECLasso}
\alias{ECLasso-package}
\title{ECLasso: Solving Equality-Constrained Generalized Linear Models or Cox Regression with Lasso Penalty via Alternating Direction Method of Multipliers (ADMM)}
\description{
An inplementation of the Alternating Direction Method of Multipliers in R.  Given data from 
    survival (Cox) certain GLMs (Gaussian, poisson, binomial), it can estimate parameters or parameter pairs for each covariate.  
    It includes options on whether to include an intercept effect, to solve exactly or use an inexact update
    in each iteration (which greatly speeds up computation time), or solve over a user-defined lambda sequence.
    At each iteration, it outputs the parameter estimates and the lambda at that iteration, as well as the
    solution path upon completion.
}
\details{
The only function you will actually run from ECLasso is ECLasso.fit.
Please refer to the ECLasso.fit documentation for more details about function options and a short example.
}
\references{
\itemize{
\item ECLasso: Fitting equality-constrained,L1-penalized models with ADMM to find gene pairs
\item \url{https://github.com/umich-biostatistics/genepADMM}
}
}
\author{
\itemize {
\item Lam Tran <lamtran@umich.edu>
\item Hui Jiang <jianghui@umich.edu>
}
}
