% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECLasso.R
\name{ECLasso}
\alias{ECLasso}
\title{Function for solving constrained lasso}
\usage{
ECLasso(
  x,
  y,
  censor = NULL,
  C = NULL,
  d = NULL,
  family = c("cox", "logistic"),
  penalty.factor = 1,
  depth = 15,
  tolerance = 1e-07,
  trace = F
)
}
\arguments{
\item{x}{The design matrix}

\item{y}{The response vector}

\item{censor}{If using a Cox model, the censoring vector}

\item{C}{The constraint that linearly combines the regression coefficients. Should be in a 1xp matrix}

\item{d}{The scalar value that C times the regrssion coefficients should equal}

\item{family}{The type of model. Note for now only logistic and Cox models are supported.}

\item{penalty.factor}{The penalization vector that multiplies the lasso penalty lambda e.g. elements = 1
are lasso penalized and elements = 0 are unpenalized}

\item{depth}{How many lambdas into the unconstrained sequence to solve for}

\item{tolerance}{Tolerance of ADMM, influences the primal and dual criteria for termination}

\item{trace}{If false, returns regression coefficients. If true, also returns number of ADMM iterations required}
}
\description{
This function solves the constrained lasso using our candidate subset approach
}
\keyword{ECLasso}
