% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvingfcts.R
\name{solvingfcts}
\alias{solvingfcts}
\alias{fast_constrained}
\alias{fast_no_nr}
\alias{naive_fit}
\title{Solvers of the constrained lasso, called by the main ECLasso function}
\usage{
fast_constrained(
  x,
  y,
  censor = NULL,
  C = NULL,
  d = NULL,
  family = c("cox", "logistic"),
  penalty.factor = 1,
  depth = 15,
  tolerance = 1e-07,
  intercept = F,
  trace = F
)

fast_no_nr(
  x,
  y,
  censor = NULL,
  C = NULL,
  d = NULL,
  family = c("cox", "logistic"),
  penalty.factor = 1,
  depth = 15,
  tolerance = 1e-07,
  intercept = F,
  trace = F
)

naive_fit(
  x,
  y,
  censor = NULL,
  C = NULL,
  d = NULL,
  family = c("cox", "logistic"),
  penalty.factor = 1,
  depth = 15,
  tolerance = 1e-07,
  intercept = F,
  trace = F
)
}
\arguments{
\item{x}{The design matrix}

\item{y}{The response vector}

\item{censor}{If using a Cox model, the censoring vector}

\item{C}{The constraints that linearly combine the regression coefficients.}

\item{d}{The vector of values that C times the regrssion coefficients should equal}

\item{family}{The type of model. Note for now only logistic and Cox models are supported.}

\item{penalty.factor}{The penalization vector that multiplies the lasso penalty lambda e.g. elements = 1
are lasso penalized and elements = 0 are unpenalized}

\item{depth}{How many lambdas into the unconstrained sequence to solve for}

\item{tolerance}{Tolerance of ADMM, influences the primal and dual criteria for termination}

\item{trace}{If false, returns solution path. If true, also returns run-time, subsets, and penalty sequence}
}
\description{
Solvers of the constrained lasso, called by the main ECLasso function
}
